Couple of <algorithm> added and lots of int casts of .size and void(*)s added from 6e7a65f02177e693af56737351f1469169f5edb4

+++ src/ipm/ipx/src/diagonal_precond.cc
@@ -55,8 +55,8 @@ void DiagonalPrecond::_Apply(const Vector& rhs, Vector& lhs,
     Timer timer;
 
     assert(factorized_);
-    assert(lhs.size() == m);
-    assert(rhs.size() == m);
+    assert((int)lhs.size() == m);
+    assert((int)rhs.size() == m);
 
     for (Int i = 0; i < m; i++) {
         lhs[i] = rhs[i] / diagonal_[i];
